Set SD for simulation = SE
In this particular instance, I think we will use the SE for simulating parameter distributions. Sample sizes are difficult to properly quantify for complex hierarchical models (like SCR models), and the SE estimates you get are very big.
We should discuss this with Matt, but for now I suggest we use the SEs for simulating parameters.