Deleted "scenarios"
I am not sure what you were trying to do here but I think you mixed up things. For running the PVA scenarios, we do not use the same setup as for sensitivity analyses. Instead, we use the kind of setup that you coded previously (in the script "FishingCat_PVA_Simulations.R").